Challenge: Box office hits database
In the "Box office hits database" challenge in Khan Academy's "Intro to SQL: Querying and Managing Data" lesson, I was challenged to do the following:

"This database contains an incomplete list of box office hits and their release year. In this challenge, you're going to get the results back out of the database in different ways! In this first step, just select all the movies."